component_trait

Trait Component

pub(crate) trait Component {
    // Required method
    fn render(&mut self, frame: &mut Frame<'_>, area: Rect);
}

A custom trait for components that can render themselves while mutating their state.

This trait provides an alternative to the standard Widget trait by allowing components to take &mut self, enabling direct state mutation during rendering.

Required Methods

fn render(&mut self, frame: &mut Frame<'_>, area: Rect)

Render the component to the given area of the frame.